usb: split mouse into grid_usb_mouse.c/h
Mouse has no tx queue — only button/move action functions called by the
macro sequencer. Move struct, state, init, on_connect/disconnect, and
the two action functions to their own unit. grid_usb_hid uses a forward
declaration for struct grid_mouse_model since it only holds a pointer.
Mouse param removed from grid_usb_hid_on_connect/disconnect.
